Read leaves, write insert, and iterator get.
Fixed read leaf to use the position after the new line.

Added some thoughts on the file format and whether or not it can be human
edited. (It can't.) Added a thought about hashing for integrity.

Created an insert test.

Write insert was not returning the insert position.

The range check for iterator was wrong, had gotten stale.

Leaf read