Implemented JSON size.
Correctly implemented JSON size accounting for every record and key cached, as
well as the length of the reference arrays.

Created a method called `heft` that adjusts the page size and b-tree size. It is
a doodle, two lines of code, but better than leaving those two lines out there,
repeated all over the code, so common they are invisible. A perfect place for a
bug to hide; in plain sight.

Created `splice` to add and remove references to a reference array. A reference
array is an addresses array in a branch page, or a positions array in leaf page.

The `splice` method is for the sake of keeping track of the JSON size. It is
easy enough to `splice` and `push` the array directly. This method ensures that
we don't neglect to appropriately adjust the JSON size of the page and b-tree.

Moving the details of the reference manipulation into a single function
simplifies the calculation of the JSON size of the reference array. So much so
that I've decided to continue to use a string length value, instead of the
product of the reference array length and an arbitrarily chose value that I was
considering.

The actual value of a number is probably larger than the character count of its
string representation, but this is a relative value. The JSON string length of
the reference array is consistent with how me measure the heft of cached records
and keys. Saves a burden of explanation.

Renamed `IO.rewriteLeaves` to `IO.rewriteLeaf`.

To ensure that JSON size is working correctly, I added a JSON size assertion to
`lock`, where I assert the JSON size of a page as it is loaded. Since all
adjustments go through `heft`, I can be reasonably certain that the b-tree size
is if I never detect problems with the pages I visit.

This is temporary code, which we'll remove when confidence in JSON page size is
profound.

The `lock` function is now relinking requested pages to the head of the
most-recently used list.

All of our `IO` read and write methods have been rewritten to correctly adjust
the heft of the cache using `cacheKey`, `uncacheKey`, `cacheRecord`,
`uncacheRecord`, and `splice`. These in turn call `heft`, while manipulating the
`addresses` or `positions` array, and the `cache` of the page.

Additionally, I added a `travis-ci` branch to `.travis.yml`. As it is
configured, Travis CI ignores pushes to all branches except for the `javascript`
branch. With the `travis-ci` branch, I can push to Travis CI before I merge to
main branch of development and push there.