Implement concurrent page load.
Multiple descents may have detected that a page has not yet been loaded
simultaneously. If a descent visited a page that was not yet loaded, it would
load it. While that was going on, another descent might visit the same unloaded
page. It would also load it.

Duplicating the reads was probably benign, probably would not not have cased
errors. If the pages were read locked, then they'd replace the page contents,
but not undo any changes, because no changes could be made, because of the read
lock. If the page was write locked, then there would only be one load anyway.

There is now a `load` queue property for an unloaded page. The presence of the
`load` queue indicates that the page is unloaded. The Docco describes the
operation. The first descent through pushes its callback onto the `load` queue,
then invokes the page read function with a callback that will invoke all the
callbacks in the `load` queue. Subsequent descents push their callbacks onto the
`load` queue and do nothing more.

No more `loaded` flag. It is the lack of a `load` queue that indicates that a
page is loaded.

I cleaned up the preamble to locking while writing the Docco for loading.

Renamed `IO.readBranches` to `IO.readBranch`. The `IO.load` function is the only
function that calls `IO.readBranch`, so it seemed like a good time to bring it
in line with `IO.readLeaf`.

Went ahead and renamed `IO.rewriteBranches` to `IO.rewriteBranch` as long as
we're working on consistency.