Accept `null` keys, warn against them.
The b&#x2011;tree can accept `null` keys in theory, but the b&#x2011;tree cannot
accept duplicates. There is no sense in a data set with `null` keys without
duplicate `null` keys.

I've updated the code to support a single `null` key in a b&#x2011;tree, while
at the same time updating the Docco to explain how silly that is.

The `Strata.iterator` and `Strata.mutator` have an optional `key` argument, the
absence of which indicates that the cursor should be positioned at the first
record in the entire b&#x2011;tree. I replaced the existence check for the key
with a check against the length of a CoffeeScript splat. This allows the key to
be `null` when either `Strata.iterator` or `Strata.mutator` are called.

That appears to be all that was necessary to support `null` keys. The value
`null` has a valid JSON representation.

Started to talk about pseudo-duplicate keys, but I've not documented them here
yet. I might document them in a cookbook and link to that cookbook recipe from
the Docco. The `Iterator` and `Mutator` readily support a method of generating a
pseudo-duplicate that is a combination of a key and a series value.

The pseudo-duplicate strategy was part of the intent of the design of those
classes, but maybe, instead of explaining those design thoughts in the Docco, I
can cookbook the pseudo-duplicate strategy, without belaboring the point that it
was deliberate, not an accident of the design.