Add test cleanup functions.
Updated for test framework rename. Also, using a new framework facility for set
up and tear down of tests. The cleanup functions in the test harnesses are run
once at the beginning of a test and once at the end. We can also specify a setup
function, which is run after the defensive cleanup at the start of the test.

This means we can cleanup our temporary directory, deleting it entirely when we
have a successful run, so we no longer to exclude these temporary directories
from `git` operations using `.gitignore`. If the test are passing, then the
temporary directories are all cleaned away.