Update for Streamline 0.3.0.
Updated for the latest Streamline by changing file names to use the `._coffee`
filename ending. Updated Twinkie to the Streamline 0.3.0 aware version of
Twinkie.

Fixed `.travis.yml` to build a new `coffeescript` branch. I'd named the main
branch `javascript` because I've always imagined writing a C or C++ version of
this library. Now, I'm probably more like to have a go at it in JavaScript with
functional composition, so I ought to move `javascript` out of the way for that
unlikely future.