Implement merge leaves.
Implemented merge leaves. Worked through the outline code. Made it run.

Removed the outlined attempts to determine the next course of action,
whether or not to proceed with filling the root or splitting a parent
branch page. The outlined logic was naive, plus I don't want to think
about this at this commit.

There was some confusion in leaf pages as to whether the right sibling
address of the right most page was `-1` or `0`. With this commit I've
gotten all the tests and tested branches of code to agree that a
non-existent right sibling is `0`.

Created a `stringify._coffee` utility to convert a temporary directory
containing a b-tree into JSON.

Adjusted the file rewriting functions to accept a suffix with the dot
separator. `mergeLeaves` needs to rename the outgoing right leaf page in
a merge from a file with no extension to a file with the extension
`".unlink"`. If the suffix includes the dot separator, then suffix can
simply be appended to the file name in the implementation of those
functions. Plus, the presence of the dot separator suggests the
intention of those functions.

Having a problem naming `IO.unlink`. I want to use it for both the
most-recently used list in the cache, and to remove a file from the file
system. I've moved the most-recently used cache `IO.unlink` to
`IO._unlink` for now.

Fixed `Descent` so that will not unlock the parent page after the first
dive after a `Descent.fork`. The forked descent does not own the parent.

Fixed `Mutator.delete`, which wasn't working even a little bit. It was
outline code. I had not written a basic test for `Mutator.delete`.

I've created a method called `hit` which I'm going to use for minmalist
coverage and tracing. I'm looking at how to use DTrace, SystemTap and
JSCoverage, but all of those look like rabbit holes at this point. I
refuse to allow this project to send me, once again, into the land of
tools, especially build tools.

Unable to create tests to merge empty pages, because I've not yet
implemented ghosted key records. We'll revisit merging after ghosted key
records are complete.