Implement ghosted records.
Cleaned up most of the tests. Many tests were incorrectly using
`Iterator.index` as start of the iterator range. This has been changed
to `Iterator.offset`. It didn't matter until ghosts started to appear.

Also, went through all of the JSON fixtures and changed the right
sibling page to `0` where it was `-1`.

Rewrote `Strata._cursor` to use the index into the leaf page generated
by the `Descent` object, which required the `Descent` object to be a
little more discerning about the page types, whether or not they were
leaf pages or branch pages. The `key` function generator now generates a
function that will use the `page.ghosts` low index in its call to `find`
when the page is a leaf page. The `descend` function does not logically
and its `index` property if the page is a leaf page. The `leftMost`
function returns `ghosts` if the page is a leaf page.

When recording a delete, we don't need a ghosts flag. We'll know. But,
we won't know who's a ghost in the positions array, so we've added
ghostliness to that journal entry. We write out the number of ghosts in
the page along with the positions array and right sibling leaf page
address.

Loading a leaf page will increment a page's `ghosts` property if the
first record is deleted and the page is not the left-most leaf page.

Calling insert objects journal entires now in the documentation.

The `Balancer` now accounts for ghosts when determining whether to merge
or split. It records the length of a leaf page's position array less its
ghosts as the length of the leaf page. It uses the adjusted length for
balance planning and when double checking the validity of the merge and
split operations.

Oddly, when calling `Descent.key`  in `Strata._cursor` was passing in
the key as an array, when the actual key value was the first element of
the array. How did that ever work?

Removed the `first` property because we'll know the left most leaf page
by its address; `-1`.

In order to perform a merge of leaf pages where the left leaf page has
ghosts, we need to first remove the ghosts. Removing ghosts requires
exclusively locking the branch page that uses the leaf's key as a pivot,
because by deleting the ghost, we're going to change the leaf's key. We
remove the cached key from the branch page so that the next decent to
visit the page will look up the new leaf page key.

We could implement `mergeLeaves` to account for a left leaf page with
ghosts, but the separation of operations makes the b&#x2011;tree easier
to read and maintain. We've long ago decided that that descents along a
cached path are too cheap for us to seek to fold them into each other,
and the folding obfuscates concurrent operations.

Deleting ghosts does not trigger a rewrite. Instead appending a position
array entry to the leaf page after removing the first position and
setting ghosts to zero, with the branch locked exclusively.

Closes #15.
Closes #13.
Closes #16.
Closes #17.