Checksums.
We not store a checksum of each line of JSON in branch page file and
leaf page files. Until we implement a CRC 32, we're using SHA1, which is
available through the `crypto` Node.js module.

The checksum is appended as a hexadecimal number after the JSON string
on each line of a branch page file or a leaf page file. Closes #9.

The test fixtures have been updated to read and write b&#x2011;tree file
directories containing branch page files and leaf page files with
checksummed lines of JSON.

Each leaf page entry now includes an entry number. This can be used to
verify the integrity of the file, to insure that no entries are missing.
The entry number should increase by one with each entry. A break in the
series indicates corruption. Closes #25.

We added this because we checksum each line of a leaf page file, but we
do not checksum the entire leaf page file. We know that each line is
correct, but without the entry count, we don't know if any lines or
missing, repeated, or out of order.

Each insert and delete entry now includes the record count of the leaf
page including the effect of the entry. This was added in anticipation
of counted our b&#x2011;tree implementation described in #10.
Closes #26.

Each position array entry now includes a leaf page file format version
number, because changes to file format are inevitable. Current version
number is `1`. Closes #28.

A position array entry is always the first entry of a leaf page file.
We'll be able to look at the head of the file to find the right sibling
leaf page. This is part of a future counted b&#x2011;tree strategy of
skimming leaf pages at startup to read leaf page record counts.
Closes #29.

Removed a long chunk of exposition on the editablity of leaf page files.
They're not really meant for hacking with `vim`. They can be managed
with `git`, viewed with `less`, watched with `tail -f`, and munged using
a Node.js program of the adopter's design. We store file positions
however, so making an edit in your favorite text editor is likely to
throw those file positions off.

Part of the value of JSON and plain text is the visibility. Part of the
value is that an application developer can restore corrupt files using
utilities written in response to a specific instance of corruption,
informed by the visibility. The data is still there. You can read
through a corrupted file, recovering what can be recovered, inserting it
into a new b&#x2011;tree.

Also, I noticed that I was not adding the dot to the file suffixes when
rewriting files in create. I added the dots.