Renamed `IO.rewriteBranch` to `IO.writeBranch`.
Branch page files are always written out in their entirety. They contain
only a single line of JSON with a checksum. There's no real rewriting as
there is with leaf page files, where we rewrite the file in order to
cleanup records that have been deleted and the journal entries that
deleted them.

The name `writeBranch` has less going on than `rewriteBranch` and there
is less going on.

Closes #27.