Using latest Stencil.
Stencil is transforming into a command line program, or a library with
the a command line invocation option. Edify is no longer a stand alone
executable for generating GitHub:Pages. Instead, it is a library that
provides transformations of Markdown and source code for Stencil, so
that Stencil can be used to generate GitHub:Pages.

Strata is a test bed for the new Stencil invocation. Currently, we're
using a `Makefile` to invoke Stencil, but we do want Stencil to be able
to build using a `make` like syntax of its own.