`code` submodule referenced missing commit.
The `code` submodule was referencing a commit that did not exist on
GitHub. I deleted the `code` directory and checked it out using the
default `coffeescript` branch, then committed the `gh-pages` branch with
the new submodule. When the GitHub:Pages built, it occurred to me that
GitHub was previously unable to checkout the submodule.

Deleting the `code` directory was impetuous and may have resulted in
lost work. I don't believe there was any meaningful work in there, but
now, of course, I'll never know.