Convert `t/basics/open.t`.
Converted and passing. The `stats` function should be an accessor.
Perhaps there ought to be accessors for `size` and `nextAddress`, but if
we're going to inspect internal state, then an object with copied
properties probably works. It's a fiddly enough to indicate that it's
not core to the interface.

Updated Proof to 0.0.7.