DRY up first test.
Upgraded Proof to 0.0.9 which allows nested control flow. The first test
was repeating itself, requesting the same objects and function in each
step of the series of set functions. With nested control flow, the
commonly used objects and variables are specified in the outer most
function, and the inner functions only request the properties that they
alone need.

Also, Proof 0.0.9 detects variables that leak into the global namespace,
so I cleaned up some leaky variables.