Go right first to find parents of pages to merge.
First go right instead of left to find the parent branch page of the
right sibling leaf page of a merge before finding the parent branch page
of the left sibling leaf page of a merge.

We're going to need to keep track of the ancestor branch pages of the
left leaf page of a leaf page merge that have only one child. If a
branch page has only one child, when that child is removed due to the
merge, we will no longer be able to designate a key for that branch
page. Any branch page that would result in having no children as a
result of the merge needs to be removed as part of the merge.

With this commit we take a step toward pruning the tree in this way by
descending to the right before the left, so that we can take note of any
branch pages that would end up with no children. With a list of branch
pages to prune, we can to left and gather a list of their left sibling
branch pages so that we can reassign the right sibling branch page
reference of those pages.

The complexity of maintaining the links between branch page siblings
appears to be greater than the benefits to navigation. The only time
we'll use them is to find the siblings of a branch page to determine if
the branch page can merge with a sibling branch page. The descent to
determine if a branch page can merge precedes a separate descent to
perform the merge.

The preliminary descent does not lock the tree exclusively, nor does it
contend with any descents that would lock a branch page exclusively. The
preliminary descent will therefore not impede the progress of any other
descent. It can hold onto its locks and dive into the tree at any branch
page. If it needs to start from the stop and work its way back down it
can dive into the tree at any point, knowing that it won't deadlock.

In the case where there will be no other descent obtaining an exclusive
lock, shared locks are reentrant, which might simplify the
implementation of the logic to locate the right sibling of a branch
page.