Swappable `unlocker` function in `Descent`.
Added an `unlocker` property to the `Descent` class. The `unlocker`
property is a function that determines whether or not to unlock pages
after they've been visited by the `Descent`. Our merge function will
provide it's own implementation of the `unlocker` function that will
gather branch pages that will be emptied by the merge, holding their
locks.