Delete branch pages empty after leaf merge.
For the leaf page deleted by a leaf page merge, the `mergeLeaves`
function now tracks which ancestor branch pages that would be left empty
when by the deletion. Any branch page on the path to the leaf page with
a single child whose descendants consist only of the leaf page or a
branch page with a single child must be deleted.

After the merge, a branch page such as this will have no leaf page
descendant from which to obtain a key. Somewhere in the b-tree there
will be a branch page with a child for which no key can be designated.

After we find the pivot branch page, we gather up branch pages that
have only one child. We discard the branch pages we've gathered the
moment we encounter a branch page with more than one child. We unlink
all of the gathered pages, if any, as part of the rewrite. We remove the
child on the path to the leaf page from youngest ancestor that has more
than one child. Balancing branch pages continues with the youngest
ancestor that had more than one child.

At this point, however, we're not performing the child removal on the
correct ancestor. The `Descent` class needs to be extended to track the
index used to descend the tree for each branch page so that we know
which child to remove. Currently, we're using the existing logic which
assumes that the penultimate branch page is the page from which a child
is removed to accommodate the merge.