Remove at ancestor, track indexes on descent.
During a merge we now remove the child on the path to the leaf page from
youngest ancestor that has more than one child. In order to know which
child to remove, the `Descent` class now tracks the index of the child
selected by the navigation function for each branch page or leaf page
visited during the descent. This map of page addresses to child indexes
is preserved across forks.