Purge cached keys on ghost delete and merge.
Now purging cached keys on the descent to delete a ghosted key or delete
a page by merging it into its left sibling. Just as this has been
implemented, it occurs to me that `designate` is the way it is so that
it will not cache a key when the index is zero. We probably do not need
to perform this purge.

We may want those keys to be cached when we set about merging branches.
A future commit will either reimplement `designate` to cache a key for
the zero index when descending or else remove this cache purging.

Note that, even if `designate` is implemented in terms of itself, and
therefore caches a key for the first child, which is not actually
referenced in the binary search when searching the tree, we will never
cache a zero index key for the path that leads to the left most leaf
page. No care needs to be taken to prevent this from happening, since
we'll never be inspecting these keys during a branch merge, because
we'll never go down that path on a branch merge. It would be unusual if
it managed to happen somehow.