Implement `level` stop function in `Descent`.
Replaced the `descendant` stop function with the `level` stop function
in the `Descent` class. The `level` function will stop at a specific
depth in the b-tree.

The `depth` property of the `Descent` has been exposed through a getter.

The one spot where `descendant` was being used has been updated to use
the `level` function instead.