Implement `lesser` and `greater` in `Descent`.
Created a `lesser` and `greater` property on the `Descent` class. This
property is as of yet unused. It will be used to find the left and right
siblings of a branch page that is a candidate for a merge.

The `lesser` property of a `Descent` is an instance of `Descent` that if
followed to the right to the depth of the owner `Descent` will arrive at
the left sibling of the current page of the owner `Descent`. If the
current page of a `Descent` has no left sibling, then the `lesser`
property is undefined.

The `greater` property of a `Descent` is an instance of `Descent` that
if followed to the left to the depth of the owner `Descent` will arrive
at the left sibling of the current page of the owner `Descent`. If the
current page of a `Descent` has no right sibling, then the `greater`
property is undefined.

The `greater` and `lesser` properties are only known to be meaningful
when descending using a `key` navigation function. It is still
maintained by the `Descent` when using other navigation functions, but
it will not be used nor unit tested except with the `key` navigation
function. Seems that it will work regardless.