Decided against `references` array.
Decided against renaming the addresses array of the branch page and the
positions array of the leaf page so that they both the had a references
array. The documentation became far more verbose as it set out to
describe the meaning of the hopelessly generic term reference.

I documented the decision and removed the todos suggesting the change.