Propagate leaf merge to branches.
Added logic to determine if a merge of leaf pages should trigger an
attempt to merge parent branch pages or fill the root. The logic
references two unimplemented functions. None of the current tests
trigger a merge that ought to propagate, so these unimplemented
functions are not invoked by the tests.