Rename `address` stop function to `child`.
The `address` function stops when it reaches a branch page that has a
certain address in its address array. I've renamed that function `child`
to distinguish it from `address`, the function that I'd expect `address`
to be, a function that stops when it reaches a specific page identified
by the address of the page.