Merge branch pages.
Implemented the merge of branch pages using the generalized merge
algorithm that, with some specializations, also merges leaf pages.

Implemented a function named `chooseBranchesToMerge` that tests whether
a branch page can be merged with a sibling after a descendant has been
merged into its sibling and a child address removed from the branch
page. This function descends to the branch page to test, then uses the
`lesser` and `greater` properties of the `Descent` class to find the
left and right siblings. If the branch page can match with either
sibling, it then invokes `mergeBranches`.

The `lesser` and `greater` of `Descent` properties were broken. They
were being captured after the `depth` was increased but before the next
descent, giving a depth that was off by one. After capturing those
properties in a fork, their indexes were not being adjusted to send them
down the lesser or greater path. Both issues have been corrected.

The `mergeBranches` function is implemented using the generalized
`mergePages` function.