Fix `found` to never test first child.
Fixed functions built by the `found` descent stopper function builder so
that they will never test for the sought key when the current index is
zero.

The `found` descent stopper stops when the we've reached the branch page
that contains a given key. This function must be used with the `key`
navigation function because it assumes that if the key is found in the
branch page, it will be at the index resolved by the navigation
function. Because the `key` navigation function will cache the key in
the cache of the branch page, the key is read directly from the cache.
We do not need to look it up.

We do not want to perform the `found` test when the index resolved by
the `key` function is zero. This indicates the first child, which is
really the left node of the first key, the key associated with the
second child. We do not read a key into the cache for the first child.

Because there is no key in the cache, the `found` function was comparing
`undefined` to the key sought and the comparator function used in our
tests, which as not designed to accommodate `undefined`, was incorrectly
returning `true` causing a descent to stop whenever it had to go to the
far left child.

We've fixed this by having the `found` function return false if the
current `index` is `0`.