Implement fill root.
Implemented the merge operation that fills the root when merges have
made it so that the root has only one child and that child is a branch
page. In essence, we promote the only branch page child to root.

The fill root operation replaces the only child of the root with the
children of the only child. The only child is then unlinked.

The fill root operation is the operation that decreases the height of
the b-tree.

Closes #7.