Implement cache purge.
Pages are purged from the cache in attempt to reduce the JSON size of
the b-tree. JSON size is an arbitrary measure of in memory heft based on
the length of the JSON strings that represent the loaded references and
objects.

If a page is not locked, but it is held by the balancer, it will not be
removed from the cache, but the page object will be reset to an unloaded
state; references and cached keys or records are discarded.

Added cache purge tests to some of the tests in the `t/basics`
directory.

Removed the `stats` property from the public `Strata` object. Replaced
it with a `size` property and a `nextAddress` property. The exposure
`nextAddress` property is dubious. It is only used in testing.

Closes #35.
Closes #12.