Document intent of balancer planning.
Not implemented yet, but added Docco to describe how the balancer
planning is supposed to work. Currently, the planner will combine more
than two pages into a merge if they will fit. I want to do them two at a
time, checking to see if further balance is possible in a subsequent
balancer.