Rename `unlink` to `terminate`.
Renamed the `unlink` function which only breaks the right link to
`terminate`. Created a new `unlink` that breaks both the left and right
links of the given node.

Note that this is the linked list of candidate leaf pages, not the leaf
pages themselves. There is another `link` and `_unlink` pair for the
pages used to maintain the most-recently used list for the cache.