Remove duplicate test for cached balance node.
Remove a duplicate test to determine if the node for the right sibling
page of a leaf page that needs to be tested for merge has already been
created. Add a test to ensure that the balancer will an existing
balancer node for the right sibling node of a merge if one has already
been created.