Implement tree structure export.
Export the tree structure for the sake of visualizations. Currently
exports the entire tree. Might actually be scalable if on provides a key
and depth parameter to expand the tree at a particular key for a
particular depth.

At this point, however, the `vivify` function is only going to be used
to print out small tree on the console, in order to create trees whose
structure tests specific edge cases in the `balancer` function.

Closes #90.