Utilities for an abstract fixture format.
Created an abstracted fixture format that is not bound to explicit file
identifiers and file positions. The abstracted fixture format records
the log entires of a leaf pages and the structure of the branch pages.
We still use file numbers for branch pages, but we normalize them for
comparison.

See #130.