Append position of last position array.
Append the position of the last position array to every entry in the
leaf page. This expedites the opening of a leaf page, no more scanning
backwards to find the last position array. It also means no more reading
backwards what-so-ever, so we can implement reading using streams.

Instead of writing out the position on the line outside of the JSON
record, I decided to write it as the last element of the JSON record.

The position can be found with a sip of the end of the file. A regular
expression can extract the final element of the JSON record. The JSON
record is always an array, so it is ordered, and the position is always
at the end.

Closes #125.
Closes #124.